ft(user-signin): refactor of code base after WIP P feedback - remove versioning from api route - change loginController.js to authcontroller.js - change loginServices.js to authServices.js [Fixes #167730631]
